requirements.txt
python-dotenv
fastapi[standard]
gunicorn>=21.2.0
jinja2
numpy<2.0.0
openai
pinecone-client
pydantic
pymongo
python-multipart
sentence-transformers
tiktoken
uvicorn